@@ -10,25 +10,25 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.1.1",
-    "@angular/common": "^16.1.1",
-    "@angular/compiler": "^16.1.1",
-    "@angular/core": "^16.1.1",
+    "@angular/animations": "^17.0.8",
+    "@angular/common": "^17.0.8",
+    "@angular/compiler": "^17.0.8",
+    "@angular/core": "^17.0.8",
     "@angular/flex-layout": "^15.0.0-beta.42",
-    "@angular/forms": "^16.1.1",
-    "@angular/platform-browser": "^16.1.1",
-    "@angular/platform-browser-dynamic": "^16.1.1",
-    "@angular/router": "^16.1.1",
+    "@angular/forms": "^17.0.8",
+    "@angular/platform-browser": "^17.0.8",
+    "@angular/platform-browser-dynamic": "^17.0.8",
+    "@angular/router": "^17.0.8",
     "@microsoft/signalr": "^7.0.7",
     "bulma": "^0.9.4",
     "rxjs": "^7.8.1",
     "tslib": "^2.3.0",
-    "zone.js": "~0.13.1"
+    "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.1.0",
-    "@angular/cli": "~16.1.0",
-    "@angular/compiler-cli": "^16.1.1",
+    "@angular-devkit/build-angular": "^17.0.8",
+    "@angular/cli": "~17.0.8",
+    "@angular/compiler-cli": "^17.0.8",
     "@types/jasmine": "^4.3.3",
     "jasmine-core": "^5.0.1",
     "karma": "^6.4.2",
@@ -36,6 +36,6 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "^5.1.0",
     "karma-jasmine-html-reporter": "^2.1.0",
-    "typescript": "~4.9.5"
+    "typescript": "~5.2.2"
   }
 }
